build: Use include files to project package type
The build will use a project's packaging type to include it into the
proper build artifacts.

The packaging include files also allow use to build different
packaging types with different attributes such as versions. This will
be used to build annotation and core as final while cmpn and promise are
still snapshot.